+++ R/draw-gam.R
@@ -22,6 +22,10 @@
 #'   [mgcv::plot.gam()], is to not draw parametric effects.
 #' @param terms character; which model parametric terms should be drawn? The
 #'   Default of `NULL` will plot all parametric terms that can be drawn.
+#'   Interactions show their individual contributions to the linear predictor;
+#'   main effects and intercepts are not added. Multivariate parametric terms
+#'   use the grid and surface controls below, as described in
+#'   [parametric_effects()].
 #' @param residuals logical; should partial residuals for a smooth be drawn?
 #'   Ignored for anything but a simple univariate smooth.
 #' @param scales character; should all univariate smooths be plotted with the